pyproject.toml
[tool.poetry]
name = "hespi"
version = "0.2.3"
description = "HErbarium Specimen sheet PIpeline"
authors = ["Robert Turnbull <robert.turnbull@unimelb.edu.au>"]
license = "Apache-2.0"
readme = "README.rst"
exclude = ["hespi/data/plants.csv"]
homepage = "https://rbturnbull.github.io/hespi/"
repository = "https://github.com/rbturnbull/hespi"
documentation = "https://rbturnbull.github.io/hespi/"
keywords = ["herbarium", "object detection", "OCR", "HTR", "specimen", "handwritten text recognition"]
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Scientific/Engineering :: Image Recognition"
]

[tool.poetry.dependencies]
python = ">=3.8,<3.12"
rich = "^10.16.1"
typer = "^0.4.0"
yolov5 = "^6.1.8"
torchapp = "^0.2.0"
pytesseract = "^0.3.10"
transformers = "^4.21.3"
appdirs = "^1.4.4"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
ipykernel = "^6.6.1"
coverage = "^5.5"
autopep8 = "^1.5.7"
Sphinx = "^4.2.0"
nbsphinx = "^0.8.7"
sphinx-rtd-theme = "^1.0.0"
sphinx-autobuild = "^2021.3.14"
myst-parser = "^0.15.2"
pre-commit = "^2.15.0"
sphinx-copybutton = "^0.4.0"
black = "^21.10b0"
sphinx-click = {git = "https://github.com/rbturnbull/sphinx-click.git"}

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = [
    "tests",
]

[build-system]
requires = ["setuptools","poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
hespi = "hespi.main:app"
